Removed sdk common directory, moved files up one level
