+++ src/mbed_cloud/client/client.py
@@ -3,7 +3,7 @@ import json
 
 from mbed_cloud import utils
 from mbed_cloud.sdk import ApiErrorResponse
-from mbed_cloud.sdk.common.config import Config
+from mbed_cloud.sdk.config import Config
 from mbed_cloud.client import util
 
 import requests
