+++ src/mbed_cloud/sdk/exceptions.py
