+++ tests/integration/server.py
@@ -41,7 +41,7 @@ from mbed_cloud import __version__
 from mbed_cloud import pagination
 from mbed_cloud.core import BaseObject
 
-from mbed_cloud.sdk.common.exceptions import ApiErrorResponse
+from mbed_cloud.sdk.exceptions import ApiErrorResponse
 from mbed_cloud.sdk.entities import __all__ as entity_list
 from mbed_cloud.sdk import SDK
